pyproject.toml
[project]
name = "Atv-Integrada-Hotelaria"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "mateusLimadeMacedo", email = "mateuslimamacedo2008@gmail.com"},
    {name = "", email = ""},
]
dependencies = [
    "django>=5.2.4",
    "django-rest-framework>=0.1.0",
    "djangorestframework>=3.16.0",
]
requires-python = ">=3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
dev = "python manage.py runserver"
